+++ NEWS.md
@@ -1,3 +1,9 @@
+## 1.3.0 SNAPSHOT
+
+* The tool adapts to coming schema change in Orders for acquisition method
+* For now the tool will support both the coming and the existing schema
+* The validation logic checks fund, budget, budget expense class, and price.
+
 ## 1.2.0 2021-12-09
 
 * Configurable time zone for dates shown in UI